+++ KafkaConsumerRetry.Tests/DelayCalculators/BackOffDelayCalculatorTests/CalculateShould.cs
@@ -4,23 +4,23 @@ namespace KafkaConsumerRetry.Tests.DelayCalculators.BackOffDelayCalculatorTests;
 
 public class CalculateShould {
     [Theory]
-    [InlineData(0, 0, 0)]
-    [InlineData(0, 1, 1)]
-    [InlineData(0, 2, 2)]
-    [InlineData(-1, 0, 0)]
-    [InlineData(-1, 1, 0)]
-    [InlineData(-1, 2, 1)]
-    [InlineData(1, 0, 1)] // Time in the future. Doesn't really make sense but including it anyway
-    [InlineData(1, 1, 2)]
-    public void Return_A_Delay(int minutesAfterCurrentTime, int retryCount, int expectedDelay) {
-        IncrementalBackOffDelayCalculator calculator = new();
+    [InlineData(0, 0)]
+    [InlineData(1, 1)]
+    [InlineData(2, 2)]
+    public void Return_A_Delay(int retryCount, int expectedDelay) {
+        MultiplyingBackOffCalculator calculator = new(TimeSpan.FromMinutes(1));
+
+        var referenceTime = new DateTimeOffset(2022, 10, 09, 12, 10, 0, TimeSpan.FromHours(10.5));
 
         var consumeResult = new ConsumeResult<byte[], byte[]> {
             Message = new Message<byte[], byte[]> {
-                Timestamp = new Timestamp(DateTimeOffset.Now + TimeSpan.FromMinutes(minutesAfterCurrentTime))
+                Timestamp = new Timestamp(referenceTime)
             }
         };
-        var timeSpan = calculator.Calculate(consumeResult, retryCount);
-        timeSpan.Should().BeCloseTo(TimeSpan.FromMinutes(expectedDelay), TimeSpan.FromSeconds(1));
+        var actual = calculator.Calculate(consumeResult, retryCount);
+
+        var expected = (referenceTime + TimeSpan.FromMinutes(expectedDelay)).UtcDateTime;
+
+        actual.Should().BeCloseTo(expected, TimeSpan.FromSeconds(1));
     }
 }