+++ KafkaConsumerRetry/DelayCalculators/IDelayCalculator.cs
@@ -3,5 +3,5 @@
 namespace KafkaConsumerRetry.DelayCalculators;
 
 public interface IDelayCalculator {
-    TimeSpan Calculate(ConsumeResult<byte[], byte[]> consumeResult, int retryIndex);
+    DateTimeOffset Calculate(ConsumeResult<byte[], byte[]> consumeResult, int retryIndex);
 }