+++ KafkaConsumerRetry/TypeRegistrationExtensions.cs
@@ -14,7 +14,7 @@ public static class TypeRegistrationExtensions {
     /// <param name="maximumConcurrent">Maximum concurrent message processors. This applies across all consumers</param>
     /// <returns>Service collection</returns>
     public static IServiceCollection AddKafkaConsumerRetry(this IServiceCollection collection, int maximumConcurrent) {
-        return collection.AddSingleton<IDelayCalculator, IncrementalBackOffDelayCalculator>()
+        return collection.AddSingleton<IDelayCalculator, MultiplyingBackOffCalculator>()
             .AddSingleton<IProducerFactory, ProducerFactory>()
             .AddSingleton<IConsumerFactory, ConsumerFactory>()
             .AddSingleton<IRateLimiter>(_ => new RateLimiter(maximumConcurrent))